Add initial Element implementation with tests
As Opal is primarily (always?) going to be run in the browser, from
now on the core runtime will contain a default browser api that sits
directly above the native dom api. There is no requirement of jquery
or any alternative library.

The Element class wraps native element objects and has a similar and
inspired API to jquery. Prototype.js is also an influencer.

All the DOM APIs will include Ajax, Element etc and will be included
in the generated opal.js file.