Added Document.ready? method for observing ready state
The Document.ready? method is useful for registering blocks of code
to be run once the document becomes ready. This adds the default
listening event for browsers that support DOMContentReady.

This method does not currently work with browsers not supporting
this event (IE < 10), but will be added at a later stage.

No tests added for this method as a manual will need to be added
in future.