Fix bugs in setting up block arguments inside block body
Until now, if a block argument wasn't actually passed then the
variable would be null. This commit fixes that so that arguments
will be nil by default. This makes it compliant with ruby.

Secondly, this commit fixes a bug where splat args in blocks were
being incorrectly gathered which resulted in the first item being
removed from the splat array. Again, this commit fixes that bug.

Fixes 5 core specs.