Initial require() detection support inside Builder and Parser
Builder now uses require() detection to try and build all your
gem or app libs in the right order. It will find each require and
maintain this order when writing the result to the output file.

This only works for files inside the same Builder (i.e. inside the
same gem), so all other requires() are ignored. There is no runtime
Kernel#require method as it is now a no-op.