Initial commit of debug mode for builder and parser
The debug mode adds two important parts to the parser:
method_missing and const_missing. when code runs in normal mode,
it is hard coded that methods and constants referenced are
assumed to exist, as this makes opal fast.

In debug mode, every constant access as well as every method send
is wrapped to test for missing methods/constants. This is slow, but
makes things easier to debug.

This commit makes every build also build a debug version so testing
is really easy and supplied by default.