Make Array Subclasses work
Due to the way arrays work in javascript, it is not possible to
inherit from the array prototype and get the length property to
correctly keep in sync.

This fix allows array subclassing by manually adding each method
defined on a subclass to the actual array instance inside
Array#allocate. This adds some overhead to array subclasses, but
is a necessary fix.