Returns last class/module body value when is an attr_* method
The file/line comment was in the way, we need to find a better way to
print out debug info.

Maybe we should store the info (file, line, extras) and add end-line
comments when the source it's ready so they won't disturb. That would help
in adding source maps afterwards.