Fix for rescue handlers without an explicit lvar
Rescue handlers with a constant, but no lvar or ivar variable would
cause a runtime error as they were dealt with as a constant to match
against. This is now fixed, and rescue bodies are checked to ensure that
if they do not have a handler part, then one is assigned as a default.