Enable method calls to native objects via method_missing
Native objects no longer need to be wrapped, and methods can be sent
directly to them. A Native object is any object that does not implement
method_missing. The method missing dispatcher then redirects to a
native_send function in the runtime which handles the dispatch to a
native object. The result is always returned, in the exception of null
and undefined, where nil is returned. This is because methods cannot be
sent to these two objects.