Add Opal::Builder for building corelib as single file
Builder is a simple class to build the opal runtime into a single
string. This is useful as it allows opal to build its runtime without
the need of opal-sprockets. Builder just looks for all the require
triggers in the source and builds up the list of files to build
together.