Remove inline attr_accessor optimizations from parser (fixes #327)
We used to inline attr_accessor/attr_reader/attr_writer to try and be
slightly more performant, but it meant that these methods could not be
overwritten. Also, we had to try and walk the arguments, which kept
breaking when parsing in splats of attrs. So, no more inlining, and
these methods are just plain old ruby methods again.