MatchData#offset/#begin/#end will only ever support 0th element because JS
It is theoretically impossible to support these methods correctly given
how JS regular expressions engine works. I.e. there is no way to get the
index of each captured group. Faking it with substring search works in
some cases, but not in all cases. It is better to explicitly fail for
n > 0 than to give correct results *some* of the time.