Add Opal::REGEXP_START and REGEXP_END constants
These will be \A and \z in ruby mode and ^ and $ in javascript
mode, giving the regexps that use them the same semantics in
both ruby and javascript.

I think a better solution would be to have opal automatically
translate \A and \z to ^ and $, but that invites a slippery
slope of trying to convert ruby regexp syntax to javascript
regexp syntax.