Extract Opal.use_gem for use inside Opal::Builder
This allows you to use gems for specific builders instead of
modifying global state. In order to do so, it moves use_gem
from being a singleton method on Opal to being a module that
extends Opal, and including that module in Opal::Builder.

To unify the Opal singleton API and Opal::Builder instance
API, add append_paths methods to both, with Opal just using
appending to the global paths, and Opal::Builder appending
the paths to the underlying file_finder inside the path_reader.