Handle null and undefined in conditionals and logic operators
Previously using null or undefined in a conditional would result
in something like: TypeError: Cannot read property '$$is_boolean'
of undefined. Additionally, this makes using null and undefined
operate as false when used in logic operators, so code such as

  do_something if foo.JS.method && foo.JS[:prop]

will work correctly even if the method returns null/undefined or
the property is null/undefined.

There is a slight performance hit from this due to the extra
checks, but I think the more correct behavior makes this change
desirable.