Support calling javascript methods as commands
This changes the grammar to support calling javascript methods
without parentheses:

  a.JS.b 1, 2

and with blocks:

  a.JS.b { |v| ...}
  a.JS.b(1) { |v| ...}
  a.JS.b do |v| ... end
  a.JS.b(1) do |v| ... end
  a.JS.b 1 do |v| ... end

This changes the previous behavior for method calls without
parentheses.  Previously, they returned the javascript property
instead of doing a method call.  Now using .JS. always does a
method call.  If you want to return a property, use a.JS[:b]
Similarly, you can no longer set properties via a.JS.b = 1,
you now need to use a.JS[:b] = 1.

This eliminates the reduce/reduce conflicts that the addition
of property setters added.  One thing this doesn't support is
syntax like:

  foo.JS[:b] += 1

I had to specifically add support to the grammar to handle that
syntax in order to eliminate the reduce/reduce conflicts, but
now it just raises an error.  I think raising an error is the
correct behavior in this case.  This would desugar to:

  foo.JS[:b] = foo.JS[:b].+(1)

which would only make sense if foo was a javascript object with
a property b that was a ruby object.  That seems fairly unlikely
in practice, but support could probably be added to handle that
case.  I think adding such support could confuse people, since
they may expect it to operate the same as:

  `foo["b"] += 1`

To support calling javascript methods with blocks, I had to copy
the implementation of CallNode#default_compile to JsCallNode,
and then modify it.  It may be possible to refactor
CallNode#default_compile to split it up into more methods, and
reduce the duplication, but I'm not sure it is worth it.