Remove flawed method check that was preventing something? from working with struct
Methods ending in ?, !, etc. are valid but this check did not like them. In addition, it would fail silently, which isn't good. Since it was only used by the struct and was incomplete, I think removing it is the best option. it's probably better to have a built-in check inside define_method anyways. Updated to latest RubySpecs, which include https://github.com/ruby/rubyspec/pull/154 and 155