Revert "Exclude files marked as excluded in spec/rubyspec from rake PATTERN glob"
This reverts commit 8731cc67e71f867849428d7cfd0a34005e609e8d.

The original comment was:

> Exclude files marked as excluded in spec/rubyspec from rake PATTERN glob
>
> E.g. in spec/rubyspec we have:
>
>   rubyspec/core/string
>   !rubyspec/core/string/scrub_spec
>
> This means when running:
>
>   $ bundle exec rake mspec_rubyspec_node \
>   PATTERN=spec/rubyspec/core/string/*_spec.rb
>
> the script in rubyspec/core/string/scrub_spec should not be executed.

But that behavior was already supported by setting the RUBYSPEC env var:

    $ bundle exec rake mspec_rubyspec_node \
      PATTERN=spec/rubyspec/core/string/*_spec.rb \
      RUBYSPEC=true

The var name can arguably be improved and can become true by default,
but keeping it optional seems better. Especially since the commit broke the
use of PATTERN with mspec_opal_node which had all files filtered out as none
of them is part of the rubyspec whitelist.

cc @vais