Optimize Struct access
When using methods on a struct, we don't need to check whether the
struct has that as a member. If it doesn't, we'll be invoking
method_missing, not the subscript operator.

Additionally, traversing the struct's members as an array is O(n), but
checking whether the struct's data object has that key is O(1), so this
commit speeds that up, as well.