Unary minus has the precedence of... unary minus
The precedence declaration for unary plus followed by a number is not needed.
It doesn't actually help to resolve any conflicts.

The reason why unary minus was given explicit precedence is so that this:

    -1 ** 2

...will parse as `-(1 ** 2)`, not `(-1) ** 2`.

Since tUMINUS has lower precedence than tPOW, we can still accomplish the
same same end by using =tUMINUS, rather than =tLOWEST. This is more
logical.

At the same time, tweak the lexer to emit '-@NUM' rather than :tUMINUS when
appropriate. This means that this parses correctly:

    puts -1.method