Fix parsing of `method do ... end` inside lambda body
The lexer has a special state in which it lexes "do" as kDO_BLOCK rather than
kDO. It enters this state when parsing the arguments for a "command". (I am
not 100% sure, but I believe a "command" is a method call without explicit
receiver and without parentheses for arguments.)

Although this 'cmdarg' state was entered when parsing the argument list
(even an empty argument list) for a command, it wasn't cleared when entering
the body of a subsequent { } block. This meant that some constructs could not
be parsed correctly inside a { } block... unless there happened to be an
intervening construct which cleared the 'cmdarg' lexer state.

So, clear the 'cmdarg' state when entering a { } body, and restore it when
exiting.