Remove superclass param from setup_class_object()
The method was used by both boot_class_object and boot_core_class_object
In the second case the passed superclass was a function, while in the 
first case was the actual class object.

After further investigation the fact that the function was passed for 
core classes was irrelevant. In fact core classes get their $$super and
$$parent set up manually in the "Initialization" section.