Add $$prototype as a proxy for a class' prototype
The .$$prototype property is now a proxy of the actual prototype, this is
preparation for the separation of the class object from the constructor,
which being a Function cannot have its superclass changed to Class. With this
change it will be possible to target a different prototype than the JavaScript
class/constructor one.