Don't force encoding on string literals
Moving towards strict-mode string literals won't be upgraded to boxed
object when inside methods and thus they won't accept properties.

This is consistent if we consider them as frozen strings avoiding
attempts to force encoding on them. When dealing explicitly with boxed
strings we're still able to set the encoding so the compiler will now
box the string before setting it when the file is marked to have an
encoding different from the standard one.