Let Opal.find_super_dispatcher stop on stubs
The last parameter of Opal.find_super_dispatcher wasn't used. It now
accepts a boolean that will control if a stub (to support
#method_missing) or null should be returned. It also makes use of the
new Opal.send2 so that if null is returned because the method is not
found and no stub was present (e.g. if the method is only called via
__send__) it will correctly call method_missing.