Fix Number#to_i with huge number
This pull request fixes the behavior of Number#to_i to that of Math.trunc() of JavaScript.

Before the fix, Number#to_i with a huge number in exponential notation returns a small number due to parseInt() specification of JavaScript.
parseInt() converts numbers to strings and then parses it. parseInt() doesn't support exponential notation.
If parseInt() encounters a character that is not a numeral in the specified radix,
it ignores it and all succeeding characters and returns the integer value parsed up to that point.

e.g.
```ruby
 # In Opal (before the fix)
1504642339053716000000.to_s #=> "1.504642339053716e+21"
1504642339053716000000.to_i #=> 1
```

Ref.
- parseInt()
  - https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/parseInt
- Math.trunc()
  - https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Math/trunc