Opal::Cache, an optional compiler cache
First, maybe how to use it:

```ruby
OPAL_CACHE=1 bundle exec rake mspec_opal_nodejs
```

I made it an ENV variable for now, not Opal::Config, mainly because there's a plethora of ways to call Opal and almost none of them support those options. This thing only works with Opal::Builder, so it doesn't work with opal-sprockets (but it would be trivial to add if it would be wanted, after all, Sprockets has its own cache - working better or worse).

As we all know, Opal compiles things slowly. So let's take probably the biggest Opal application out there, that is, `bundle exec rake mspec_ruby_nodejs`. So on my computer it takes 1 minute and 20 second. But after cache is populated, it takes 23 seconds, almost a minute less.

Now, HOW does it work, well, this thing is a bit hacky I must say, but maybe someone will help me improve it. We cache Opal::Compiler instances (and Opal::SourceMap::File) with hand-crafted marshal_(dump,load) methods. We assume that some methods of Compiler/SourceMap won't be called, so we simply skip their instance variables if possible. Such a compiler would be useful for extracting what you usually want to extract from it, but no more.

But how and where it's stored? On disk, by 2 cache keys.

One cache key is an environment key, which is dependent on compiler options and... mtime of Opal compiler (ie. everything in lib/). The second is file key, which basically means a tuple of (filename, source, options). So you can work on Opal and have things run fast (unless you modify the compiler part, of course).

After running `mspec_ruby_nodejs` the cache is 40MB. Deflating it is certainly a good idea.

The cache also isn't pruning itself automatically. It's up to user to do that at this point. At a later time we can think about some LRU, or other method so that it may be seamlessly included as a default option.