Backport some methods from opal-rspec
In addition, modify the Struct initializer a little. In my experience,
I have never encountered using this to name a constant, yet this is
the default path. Let's try to create a shortpath.

The main reason for this, though, is that I really struggled with
the browser debugger trying to note all caught exceptions.

Co-authored-by: Elia Schito <elia@schito.me>