Don't define self = this if it's not needed
In Opal, we always alias `this` to `self`. But, let's consider the
following code. Do we really need to define `self` in that case?

```ruby
def loop
  while true
    yield
  end
end
```

And so, if a function doesn't reference self in any way (a special
case will be x-strings), let's not compile it in. So, what gains for
AsciiDoctor do we get?

```
Comparison of the Asciidoctor (a real-life Opal application) compile and run:
                  Compile time: 6.239 -> 6.123 (change: -1.86%)
                      Run time: 0.285 -> 0.284 (change: -0.10%)
                   Bundle size: 5257437 -> 5236087 (change: -0.41%)
          Minified bundle size: 1264503 -> 1254455 (change: -0.79%)
```

Not much, but at least some. Take note - we can't reliably compare the
first two performance metrics. And also gains for different softwares will be
different. The minified bundle size is the minification with `terser -c`.
Do also note, that all the following outputs of this kind will refer to entire
patchset, as compared to Opal v1.3.2.