Optimize variable creation for block passing
Ok, now for a very simple function:

```ruby
def a(&block)
end
```

You may wonder, what does this function compile to?

```javascript
  return (Opal.def(self, '$a', $a$1 = function $$a() {
    var $iter = $a$1.$$p, block = $iter || nil, self = this;

    if ($iter) $a$1.$$p = null;

    if ($iter) $a$1.$$p = null;;
    return nil;
  }, $a$1.$$arity = 0), nil) && 'a'
```

*You can see `$a$1`, which is a reference to the method body, on which
the block is attached as `$$p` when the method is called.*

Hm, not so good. One statement is duplicated, too many variable
declarations. We will focus quite a lot on this method, optimizing it
step-by-step in the following passes.

So, after some tinkering, this is the result:

```javascript
  return (Opal.def(self, '$a', $a$1 = function $$a() {
    var block = $a$1.$$p || nil;

    if (block) $a$1.$$p = null;

    ;
    return nil;
  }, $a$1.$$arity = 0), nil) && 'a'
```

Much better, right? How about the numbers?

```
Comparison of the Asciidoctor (a real-life Opal application) compile and run:
                  Compile time: 6.225 -> 6.074 (change: -2.43%)
                      Run time: 0.285 -> 0.283 (change: -0.68%)
                   Bundle size: 5257437 -> 5220199 (change: -0.71%)
          Minified bundle size: 1264503 -> 1244281 (change: -1.60%)
```

Okay! That's not bad!