Remove nil from def/defs generated code
The previous versions of Ruby tended to return `nil` for method definition, but later changed to return `Symbol`s (i.e. `Strings` in Opal). So why we
actually need this `nil`? Let's reduce it:

```javascript
  return (Opal.def(self, '$a', $a$1 = function $$a() {
    var block = $a$1.$$p || nil;

    if (block) $a$1.$$p = null;

    ;
    return nil;
  }, $a$1.$$arity = 0), 'a')
```

And the numbers:

```
Comparison of the Asciidoctor (a real-life Opal application) compile and run:
                  Compile time: 6.221 -> 6.097 (change: -1.98%)
                      Run time: 0.284 -> 0.283 (change: -0.21%)
                   Bundle size: 5257437 -> 5218997 (change: -0.73%)
          Minified bundle size: 1264503 -> 1243880 (change: -1.63%)
```

That's not much. Even though Opal has a lot of method definitions. But let's go ahead.