Optimize splice & slice calls
We use those calls a lot for extracting Ruby rest arguments from the `arguments` array-like object in JavaScript, in methods
like `def a(arg, *restargs)`. `Opal.slice` is short for `Array.prototype.slice`.

Before this step, we used to output this:

```javascript
    $post_args = Opal.slice.call(arguments, 0, arguments.length);
```

Now we output this, which is equivalent:

```javascript
    $post_args = Opal.slice.call(arguments);
```

Numbers:

```
 Comparison of the Asciidoctor (a real-life Opal application) compile and run:
                  Compile time: 6.213 -> 6.107 (change: -1.71%)
                      Run time: 0.285 -> 0.282 (change: -0.81%)
                   Bundle size: 5257437 -> 5209571 (change: -0.91%)
          Minified bundle size: 1264503 -> 1234386 (change: -2.38%)
```

So, we are accelerating.