Don't needlessly define JS vars for iter/def
Back to our `def a(&block)` empty method. Can we optimize this part even further?

```javascript
  }, $a$1.$$arity = 0), 'a')
```

Also, why do we need this, isn't it wasteful? (do note that we also need to do `var $a$1`)

```javascript
  return ($def(self, '$a', $a$1 = function $$a() {
```

Let's optimize those things out:

```javascript
Opal.queue(function(Opal) {/* Generated by Opal 1.3.1 */
  var self = Opal.top, $nesting = [], nil = Opal.nil, $$$ = Opal.$$$, $$ = Opal.$$, $def = Opal.def;

  return ($def(self, '$a', function $$a() {
    var block = $$a.$$p || nil;

    if (block) $$a.$$p = null;

    ;
    return nil;
  }, 0), 'a')
});
```

Note the `0` argument. This is a shorthand for `{$$arity: 0}`.
We can't do this optimization for a minority of methods that need
additional properties set. So, what are our gains now?

```
Comparison of the Asciidoctor (a real-life Opal application) compile and run:
                  Compile time: 6.249 -> 6.058 (change: -3.05%)
                      Run time: 0.286 -> 0.279 (change: -2.17%)
                   Bundle size: 5257437 -> 5036072 (change: -4.21%)
          Minified bundle size: 1264503 -> 1089609 (change: -13.83%)
```

That's a serious optimization now, isn't it? Almost 14% lesser files?
But let's not finish here, but go ahead, maybe we can get even smaller files!