Use `delete scope.$$p` saving one check and a few bytes
Why do we use things like `$$a.$$p`? Well - you may not know, if you aren't too familiar with Opal. This is how we pass a block argument, by setting a
`$$p` property on a called function. After a call, we unset it. But this statement: `if (block) $$a.$$p = null;`... why can't we just `delete $$a.$$p;`?
Do we lose some performance then? Perhaps - but not noticeably. And we gain a lot of space. So, our `def a(&block)` method compiled now looks like this:

```javascript
Opal.queue(function(Opal) {/* Generated by Opal 1.3.1 */
  var self = Opal.top, $nesting = [], nil = Opal.nil, $$$ = Opal.$$$, $$ = Opal.$$, $def = Opal.def;

  return ($def(self, '$a', function $$a() {
    var block = $$a.$$p || nil;

    delete $$a.$$p;

    ;
    return nil;
  }, 0), 'a')
});
```

And the numbers aren't much better, but cumulatively they are better:

```
 Comparison of the Asciidoctor (a real-life Opal application) compile and run:
                  Compile time: 6.239 -> 6.057 (change: -2.92%)
                      Run time: 0.286 -> 0.280 (change: -2.03%)
                   Bundle size: 5257437 -> 5032881 (change: -4.27%)
          Minified bundle size: 1264503 -> 1087119 (change: -14.03%)
```