Optimize toplevel constant resolution
What does `::String` compile into? Of course into `$$$("::", "String")`. Why? `"::"` is a special value here. If we would do something like `::A::B`, we would
get `$$$($$$("::", "A"), "B")`.

But why can't it become just... `$$$("String")`? It can. And our numbers now are:

```
Comparison of the Asciidoctor (a real-life Opal application) compile and run:
                  Compile time: 6.269 -> 6.195 (change: -1.18%)
                      Run time: 0.287 -> 0.279 (change: -2.75%)
                   Bundle size: 5257437 -> 5020671 (change: -4.50%)
          Minified bundle size: 1264503 -> 1077085 (change: -14.82%)
```

Not too much, but... we also needed to replace a lot of calls in the corelib from `String` to `::String`. This will follow in the next steps.