Optimize unless calls output
Ok - let's take this expression:

```ruby
true unless false
```

What will it compile to?

```javascript
  if ($truthy(false)) {
  } else {
    true
  };
```

Well - makes some sense. Oh, you may ask do we need this $truthy call
here? Well - in this particular example - we don't - but in general,
JavaScript has different
truthiness semantics. `""` is falsy, `0` is falsy, `nil` is truthy
(yeah - our `nil` is not JS `null`). But why an `if` and `else`
branch. Let's do it better:

```javascript
  if (!$truthy(false)) {
    true
  };
```

And the numbers are:

```
Comparison of the Asciidoctor (a real-life Opal application) compile and run:
                  Compile time: 6.164 -> 6.048 (change: -1.87%)
                      Run time: 0.284 -> 0.277 (change: -2.62%)
                   Bundle size: 5257437 -> 4994938 (change: -4.99%)
          Minified bundle size: 1264503 -> 1072746 (change: -15.16%)
```

Not much better in the minified bundle size - Terser did a good job
here. But in the following it didn't...