Try to eliminate closured if statements.
In case we encounter a recv or expr, let's try to create
a ternary operator. This should provide with a better
opportunity for JS minifiers.

Unfortunately, the compiler performance may tank a little bit.

---

```ruby
a || b || c || d || e
```

Seems simple, right? Can we compile it to the same thing in JS? Oh well, we can't... we have different truthy semantics as mentioned above. And also, if there's
a `next` call... you know `a || continue` is an invalid JavaScript? So... this code compiles to the following monster:

```javascript
  if ($truthy(($ret_or_1 = (function() {if ($truthy(($ret_or_2 = (function() {if ($truthy(($ret_or_3 = (function() {if ($truthy(($ret_or_4 = self.$a()))) {
    return $ret_or_4
  } else {
    return self.$b()
  }; return nil; })()))) {
    return $ret_or_3
  } else {
    return self.$c()
  }; return nil; })()))) {
    return $ret_or_2
  } else {
    return self.$d()
  }; return nil; })()))) {
    return $ret_or_1
  } else {
    return self.$e()
  }
```

Well. That's a lot of functions. And expressions like `next` don't happen here. Can't we at least use a ternary operator where we can:

```javascript
  if ($truthy(($ret_or_1 = ($truthy(($ret_or_2 = ($truthy(($ret_or_3 = ($truthy(($ret_or_4 = self.$a())) ? ($ret_or_4) : (self.$b())))) ? ($ret_or_3) : (self.$c())))) ? ($ret_or_2) : (self.$d()))))) {
    return $ret_or_1
  } else {
    return self.$e()
  }
```

This is still ugly. But we don't abuse the functions. Numbers:

```
Comparison of the Asciidoctor (a real-life Opal application) compile and run:
                  Compile time: 6.269 -> 6.198 (change: -1.13%)
                      Run time: 0.286 -> 0.273 (change: -4.44%)
                   Bundle size: 5257437 -> 4868356 (change: -7.40%)
          Minified bundle size: 1264503 -> 1069576 (change: -15.42%)
```

We improved the performance quite a bit. And the un-Tersered code size - but Terser also gained a bit. We lost a bit of compiler performance though.