Closure tracking
This adds closure awareness to Opal. Previously, some control flow
instructions were working based on the most common situation and
in certain situations we got that wrong. The previous implementation
also led to a lot of scattered around code and a lot of logic
duplication. My view is that this patch improves maintainability
of our codebase a lot and allows for further improvements based on
the new mini-DSL. Unfortunately, one of the downsides of this code
is that runtime performance is slightly decreased.