Helperization across the codebase to assist JS compressors
In compiled Asciidoctor I found a lot of dynamic regexps. And we define them by `Opal.regexp([a,b,c])`. Let's make it just `$regexp([a,b,c])` and let's shorten
a lot of other definitions like this. At this point I noticed, that we don't run Terser with name mangling in effect. Let's change it just now. The numbers are
compared to Opal 1.3.2.

```
Comparison of the Asciidoctor (a real-life Opal application) compile and run:
                  Compile time: 6.089 -> 6.006 (change: -1.36%)
                      Run time: 0.285 -> 0.269 (change: -5.39%)
                   Bundle size: 5259054 -> 4824589 (change: -8.26%)
          Minified bundle size: 1264953 -> 1054974 (change: -16.60%)
            Mangled & minified: 812275 -> 732066 (change: -9.87%)
```

That's fair enough.