Optimize self.$$prototype to $proto
For two reasons, we set `@variables` to `nil` by default if they are referenced. The first reason is obvious, `@variable` is compiled to `self.variable` and we
don't want `undefined` values to creep in - they are not an object and in Ruby everything is an object - we want to keep that impression, so in Opal `undefined`
doesn't exist (if you get `undefined` somewhere - you have hit a bug or accessed some low level interfaces). The second is to improve a shape for the JS engines
to optimize the code better.

The problem is, that the code looks like this:

```javascript
self.$$prototype.variable1 = self.$$prototype.variable2 = self.$$prototype.variable3 = self.$$prototype.variable4 = nil
```

Why not make it:

```javascript
var $proto = self.$$prototype;
$proto.variable1 = $proto.variable2 = $proto.variable3 = $proto.variable4 = nil
```

Remember - `$proto` can be safely renamed. `self.$$prototype` can't.

```
Comparison of the Asciidoctor (a real-life Opal application) compile and run:
                  Compile time: 6.086 -> 6.014 (change: -1.19%)
                      Run time: 0.284 -> 0.269 (change: -5.28%)
                   Bundle size: 5259054 -> 4823880 (change: -8.27%)
          Minified bundle size: 1264953 -> 1053776 (change: -16.69%)
            Mangled & minified: 812275 -> 730112 (change: -10.12%)
```