Decrease size of add_stubs calls
How does `#method_missing` work on Opal? In JavaScript there's no facility for that. Well - we define so-called stubs, which means that for every call you want
to make, we define a method on `BasicObject` that basically calls `#method_missing`. This way no method is missing and all calls success. And if you use a call
like `#send`... we have an easier job here, but we don't want to use `#send` everywhere for performance reasons.

The stubs used to be defined this way, for every file:

```javascript
Opal.add_stubs(["$hello", "$new", "$<"]);
```

Let's make it shorter:

```javascript
Opal.add_stubs("hello,new,<");
```

This also helps the JS parsers. This is how Google Closure Compiler optimizes large arrays of Strings.

```
 Comparison of the Asciidoctor (a real-life Opal application) compile and run:
                  Compile time: 6.088 -> 6.008 (change: -1.30%)
                      Run time: 0.285 -> 0.271 (change: -5.08%)
                   Bundle size: 5259054 -> 4812284 (change: -8.50%)
          Minified bundle size: 1264953 -> 1044964 (change: -17.39%)
            Mangled & minified: 812275 -> 721296 (change: -11.20%)
```