Don't output $$ and/or $$$ if not needed
What is `$$$` - I explained in one of the earlier parts. But what is `$$` - I haven't. This is a relative constant access function. This is a bit less performant,
because we have to iterate thru every `class` and `module` we are in and their ancestors - and `Object` and its ancestors as well. We store a list of modules and
classes in a `$nesting` variable. And then we can call `$$($nesting, "String")` to find our `String` - because - maybe it is an `Array::String`? Well, we know it
isn't, so we have to change our `corelib` furthermore a lot. And then - suddenly - some files don't need `$$`, so we don't need to helperize it.

```
Comparison of the Asciidoctor (a real-life Opal application) compile and run:
                  Compile time: 6.082 -> 5.964 (change: -1.94%)
                      Run time: 0.284 -> 0.270 (change: -4.92%)
                   Bundle size: 5259054 -> 4811653 (change: -8.51%)
          Minified bundle size: 1264953 -> 1041721 (change: -17.65%)
            Mangled & minified: 812275 -> 720161 (change: -11.34%)
```