Make $$ not need $nesting argument.
I came upon an idea, that the `$$` method can be curried. Of course, this moves its definition from the top level scope to the class/module scope so it means
it may happen more often. So now we don't call `$$($nesting, "String")`, but we can simply call `$$("String")` because `$$` is defined with `$nesting`. Do we
get any optimization from that, then?

```
 Comparison of the Asciidoctor (a real-life Opal application) compile and run:
                  Compile time: 6.097 -> 5.976 (change: -1.99%)
                      Run time: 0.285 -> 0.270 (change: -5.41%)
                   Bundle size: 5259054 -> 4795627 (change: -8.81%)
          Minified bundle size: 1264953 -> 1027837 (change: -18.75%)
            Mangled & minified: 812275 -> 716231 (change: -11.82%)
```

Yes. And quite a big one if we don't mangle variable names.