Skip empty string literal while building a dstr if first child is a str
What do we do with strings like `"aaaa#{true}"` (also called dstrs)?

Of course, we compile them to:

```javascript
"" + "aaaa" + (true)
```

Why does it make sense? Also, how comes this thing can use the `+` operator? Well, let me explain. In JavaScript, `"" + obj` is actually equivalent to
`"" + obj.toString()`. And `toString()` for Opal objects call `#to_s` - so this is exactly what `"aaaa#{true}` does in Ruby.

And you may say - ok, for strings like `"#{5}"` (being compiled to `"" + 5`) this makes sense. But if the first part of a dstr is a string, we don't need this
`""`. Yes - though, Terser applies this optimization, so there's 0 gain there.

```
 Comparison of the Asciidoctor (a real-life Opal application) compile and run:
                  Compile time: 6.087 -> 5.959 (change: -2.10%)
                      Run time: 0.284 -> 0.269 (change: -5.45%)
                   Bundle size: 5259054 -> 4783491 (change: -9.04%)
          Minified bundle size: 1264953 -> 1027837 (change: -18.75%)
            Mangled & minified: 812275 -> 716231 (change: -11.82%)
```