Also skip $parent_nesting if not needed
This is a small one. But this is the last one in this patch series. Let's conclude it with compilation of this Ruby code:

```ruby
class A
  def x
  end
end
```

Opal 1.3.2 outputs this:

```javascript
Opal.queue(function(Opal) {/* Generated by Opal 1.3.2 */
  var self = Opal.top, $nesting = [], nil = Opal.nil, $$$ = Opal.$$$, $$ = Opal.$$, $klass = Opal.klass;

  return (function($base, $super, $parent_nesting) {
    var self = $klass($base, $super, 'A');

    var $nesting = [self].concat($parent_nesting), $A_x$1;

    return (Opal.def(self, '$x', $A_x$1 = function $$x() {
      var self = this;

      return nil
    }, $A_x$1.$$arity = 0), nil) && 'x'
  })($nesting[0], null, $nesting)
});
```

This patchset makes it output the following:

```javascript
Opal.queue(function(Opal) {/* Generated by Opal 1.3.2 */
  var $nesting = [], nil = Opal.nil, $klass = Opal.klass, $def = Opal.def;

  return (function($base, $super) {
    var self = $klass($base, $super, 'A');

    return $def(self, '$x', function $$x() {

      return nil
    }, 0)
  })($nesting[0], null)
});
```

Therefore we skip one variable more. And while some JS minifiers may find this thing and optimize it out itself, some don't